# BackEnd/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"RoleResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"softDelete": {
						"type": "boolean"
					},
					"translations": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"_id",
					"name",
					"softDelete"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RoleCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"translations": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FormTemplateResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"templateGroupId": {
						"type": "string"
					},
					"template": {
						"type": "string"
					},
					"allowedSubmitRoles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"allowedSubmitUnits": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"availableFrom": {
						"type": "string",
						"format": "date-time"
					},
					"availableTo": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"title",
					"version",
					"templateGroupId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FormTemplateCreationParams": {
				"properties": {
					"template": {
						"type": "string"
					},
					"previousTemplateId": {
						"type": "string"
					}
				},
				"required": [
					"template"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "IPT Form Builder API",
		"version": "1.0.0",
		"description": "Built by GP Team B 25/26:\n- Ângela Sebastião\n- Bernardo Ferreira\n- Catarina Agostinho\n- Joana Flores\n- José Fernandes\n- Paulo Pereira\n- Sady Neto",
		"contact": {}
	},
	"paths": {
...
		"/roles": {
			"get": {
				"operationId": "GetRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/RoleResponse"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "list existing roles that are not soft-deleted (authenticated users only).",
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RoleResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					},
					"409": {
						"description": "Role already exists"
					}
				},
				"description": "add a new role (admin only).",
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoleCreationParams"
							}
						}
					}
				}
			}
		},
		"/roles/{id}": {
			"put": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RoleResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					},
					"404": {
						"description": "Role not found"
					}
				},
				"description": "update an existing role (admin only).",
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RoleCreationParams"
							}
						}
					}
				}
			}
		},
		"/roles/{id}/soft-delete": {
			"post": {
				"operationId": "SoftDeleteRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					}
				},
				"description": "soft delete a role by ID (admin only).",
				"tags": [
					"Roles"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/formTemplates": {
			"post": {
				"operationId": "CreateFormTemplate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/FormTemplateResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid JSON or missing fields"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Previous template not found"
					}
				},
				"description": "create or update a form template",
				"tags": [
					"FormTemplates"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FormTemplateCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetActiveTemplates",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FormTemplateResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "get all active form templates (dropdown data)",
				"tags": [
					"FormTemplates"
				],
				"security": [],
				"parameters": []
			}
		},
		"/formTemplates/{id}": {
			"get": {
				"operationId": "GetTemplateById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/FormTemplateResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Template not found"
					}
				},
				"description": "get a specific form template by ID",
				"tags": [
					"FormTemplates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/formTemplates/{id}/soft-delete": {
			"post": {
				"operationId": "SoftDeleteTemplate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Template not found"
					}
				},
				"description": "soft delete a form template by ID",
				"tags": [
					"FormTemplates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}